+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [BC-Break] Renamed params in global [configuration file](./config/soluble-mediatools.global.php)
 - [BC-Break] Renamed `VideoInfoService::getMediaInfo()` in `VideoInfoService::getInfo()`
 - Moved internal class `VideoInfo` in `Video\Info` namespace.
+- `VideoConversionService` set `withOverwrite()` by default.
+
 
 ### Added