+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - [BC-Break] Renamed params in global [configuration file](./config/soluble-mediatools.global.php)
 - [BC-Break] Renamed `VideoInfoService::getMediaInfo()` in `VideoInfoService::getInfo()`
+- [BC-Break] Renamed `VideoConversionService::getConversionProcess()` in `getSymfonyProcess()`
 - Moved internal class `VideoInfo` in `Video\Info` namespace.
 - `VideoConversionService` set `withOverwrite()` by default.